pyproject.toml
[project]
name = "sotaysinhvien-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "beautifulsoup4>=4.13.5",
    "dotenv>=0.9.9",
    "fastapi[all]>=0.116.1",
    "ipywidgets>=8.1.7",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-huggingface>=0.3.1",
    "langchain-pinecone>=0.2.11",
    "langgraph>=0.6.6",
    "pinecone-client>=6.0.0",
    "pypdf>=6.0.0",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]

